+++ src/main/kotlin/com/github/com/cakevm/intellij_huff_plugin/language/psi/element/HuffBuildInFnErrorCallElement.kt
@@ -1,0 +1,5 @@
+package com.github.com.cakevm.intellij_huff_plugin.language.psi.element
+
+import com.github.com.cakevm.intellij_huff_plugin.language.psi.element.base.HuffReferenceElement
+
+interface HuffBuildInFnErrorCallElement : HuffReferenceElement