+++ src/main/kotlin/com/github/com/cakevm/intellij_huff_plugin/language/reference/HuffResolver.kt
@@ -30,6 +30,9 @@ object HuffResolver {
           is HuffBuildInFnFuncSigCall -> {
             resolveUsingImports(HuffFunctionAbiDefinition::class.java, element, element.containingFile)
           }
+          is HuffBuildInFnErrorCall -> {
+            resolveUsingImports(HuffErrorDefinition::class.java, element, element.containingFile)
+          }
           else -> emptySet()
         }
       CachedValueProvider.Result.create(result, PsiModificationTracker.MODIFICATION_COUNT)